datamatrix-app/requirements.txt
flask
psycopg2-binary
python-dotenv
pandas
openpyxl
gunicorn
xlsxwriter
Flask-Login
Flask-WTF
bcrypt